(BSR)[API] chore: Move Algolia settings-related commands to its own module
We'll add other modules in `search/commands/` for other commands.
