+++ api/src/pcapi/core/search/commands/settings.py
