+++ api/tests/core/search/test_commands_settings.py
@@ -7,24 +7,24 @@ import pytest
 import requests_mock
 
 from pcapi import settings
-from pcapi.core.search import commands
+import pcapi.core.search.commands.settings as commands_settings
 
 
 class AlgoliaSettingsTest:
-    @pytest.mark.parametrize("index_type", list(commands.IndexTypes))
+    @pytest.mark.parametrize("index_type", list(commands_settings.IndexTypes))
     def test_can_get_an_index_client_from_index_name(self, index_type):
         # when
-        client = commands._get_index_client(index_type)
+        client = commands_settings._get_index_client(index_type)
 
         # then
         assert isinstance(client, SearchIndex)
         assert client.name == index_type.value
         assert client.app_id == settings.ALGOLIA_APPLICATION_ID
 
-    @pytest.mark.parametrize("index_type", list(commands.IndexTypes))
+    @pytest.mark.parametrize("index_type", list(commands_settings.IndexTypes))
     def test_can_retrieve_settings_from_algolia(self, index_type):
         # given
-        index = commands._get_index_client(index_type)
+        index = commands_settings._get_index_client(index_type)
         index_settings = {"random_field": "random_value"}
 
         with requests_mock.Mocker() as requests_mocker:
@@ -36,18 +36,18 @@ class AlgoliaSettingsTest:
             )
 
             # when
-            outputs = commands._get_settings(index)
+            outputs = commands_settings._get_settings(index)
 
         # then
         assert requests_mocker.called_once
         assert len(outputs) == 1
         assert json.dumps(index_settings, indent=4) in outputs
 
-    @pytest.mark.parametrize("index_type", list(commands.IndexTypes))
+    @pytest.mark.parametrize("index_type", list(commands_settings.IndexTypes))
     def test_can_send_settings_to_algolia(self, index_type):
         # given
-        index = commands._get_index_client(index_type)
-        config_path = commands._get_index_default_file(index_type)
+        index = commands_settings._get_index_client(index_type)
+        config_path = commands_settings._get_index_default_file(index_type)
         old_index_settings = {"random_field": "random_value"}
         config_file_content = {"random_field": "other_value"}
 
@@ -72,20 +72,20 @@ class AlgoliaSettingsTest:
             )
 
             # when
-            outputs = commands._set_settings(index, config_path, dry=False)
+            outputs = commands_settings._set_settings(index, config_path, dry=False)
 
         # then
         assert requests_mocker.call_count == 2
-        assert mock_open.call_once_with(commands._get_index_default_file(index_type), "r")
+        assert mock_open.call_once_with(commands_settings._get_index_default_file(index_type), "r")
         put_request = requests_mocker.request_history[1]
         assert put_request.text == json.dumps(config_file_content)
         assert len(outputs) == 1
         assert json.dumps(old_index_settings, indent=4) in outputs
 
-    @pytest.mark.parametrize("index_type", list(commands.IndexTypes))
+    @pytest.mark.parametrize("index_type", list(commands_settings.IndexTypes))
     def test_dry_settings_retrieval_actually_does_nothing(self, index_type):
         # given
-        index = commands._get_index_client(index_type)
+        index = commands_settings._get_index_client(index_type)
 
         with requests_mock.Mocker() as requests_mocker:
             requests_mocker.register_uri(
@@ -100,7 +100,7 @@ class AlgoliaSettingsTest:
             )
 
             # when
-            outputs = commands._get_settings(index, dry=True)
+            outputs = commands_settings._get_settings(index, dry=True)
 
         # then
         assert not requests_mocker.called
@@ -108,11 +108,11 @@ class AlgoliaSettingsTest:
         assert index_type.value in outputs[0]
         assert index_type.value in outputs[1]
 
-    @pytest.mark.parametrize("index_type", list(commands.IndexTypes))
+    @pytest.mark.parametrize("index_type", list(commands_settings.IndexTypes))
     def test_dry_settings_applying_actually_does_nothing(self, index_type):
         # given
-        index = commands._get_index_client(index_type)
-        config_path = commands._get_index_default_file(index_type)
+        index = commands_settings._get_index_client(index_type)
+        config_path = commands_settings._get_index_default_file(index_type)
 
         with (
             requests_mock.Mocker() as requests_mocker,
@@ -130,7 +130,7 @@ class AlgoliaSettingsTest:
             )
 
             # when
-            outputs = commands._set_settings(
+            outputs = commands_settings._set_settings(
                 index,
                 config_path,
                 dry=True,
